almost ready

+++ HashProduct.Domain/Models/Discount.cs
@@ -7,6 +7,6 @@ namespace HashProduct.Domain.Models
     public class Discount
     {
         public float Pct { get; set; }
-        public int ValueInCents { get; set; }
+        public float ValueInCents { get; set; }
     }
 }
