+++ HashProduct.Domain/Models/Product.cs
@@ -7,7 +7,7 @@ namespace HashProduct.Domain.Models
     public class Product
     {
         public string Id { get; set; }
-        public int PriceInCents { get; set; }
+        public float PriceInCents { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public Discount Discount { get; set; }
